+++ src/systems/si/units/mod.rs
@@ -8,8 +8,8 @@ pub use base::{Ampere, Candela, Kelvin, Kilogram, Meter, Mole, Radian, Second};
 pub use derived::{
     CubicMeter, CubicMeterPerSecond, Gray, GrayPerSecond, Hertz, Joule, JoulePerCubicMeter,
     JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin, JoulePerSquareMeter,
-    KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond, NewtonPerMeter, Pascal,
-    PascalSecond, PerKelvin, PerMeter, PerSecond, PerSquareKelvin, Rayl, SquareMeter,
+    KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond, Newton, NewtonPerMeter,
+    Pascal, PascalSecond, PerKelvin, PerMeter, PerSecond, PerSquareKelvin, Rayl, SquareMeter,
     SquareMeterPerKilogram, SquareMeterPerSecond, Watt, WattPerCubicMeter, WattPerKilogram,
     WattPerMeterKelvin, WattPerSquareMeter,
 };